Added a -o option to fasta/fastq and permit sharing filenames.
Also fixed the -F option so it defaults to 0x900 instead of that being
an internal hard-coded (undocumented) extra feature, along with
misleading docs that imply it's not done automatically.

Finally if reading from a non-terminal it'll use stdin instead of
needing an explicit "-" filename, although if present it is accepted.

Fixes #874.