Make sort and merge work with 64 bit positions
Separate tid from pos in heap1_t.  Merge rev with pos as only
one bit is needed.

Radix sort is changed to work exclusively on bytes instead of
SHIFT bits at a time.  It now counts the number of bytes needed
to store pos and tid separately, which should make it more
efficient when pos is small.